update backend uuid

+++ golang-restful-api/controllers/authCon.go
@@ -77,7 +77,7 @@ func Login(c *gin.Context) {
 	}
 
 	// Generate token JWT menggunakan helper
-	tokenString, err := utils.GenerateJWT(uint(user.Id))
+	tokenString, err := utils.GenerateJWT(user.Id)
 	if err != nil {
 		c.JSON(http.StatusUnauthorized, gin.H{"error": "Failed to generate token, Internal Server Error", "authenticated": false})
 		return
